[frontend] creation GroupConfidenceLevel(#6878)
